warn about unreconciled pages
This commit introduces a new exception type
`ConfluenceUnreconciledPageError` to help provide users a more detailed
description when they experience publishing a page which Confluence
reports is unreconciled.

This is a result of a Confluence issue observed between a 7.0 release
and a 7.3.2 release where a Confluence upgrade can result in
unreconciled pages (unknown if recent micro version of 7.0.x, 7.1.x or
7.2.x have ported a hotfix for this). More details can be found in
CONFSERVER-59196 [1].

A workaround to this issue is to have a user refresh an unreconciled
page, which is not* possible from this extension (*since we limit
configurations to use API credentials and the request needs to happen
outside the API endpoint). This should be less of a concern recently
since any user experiencing an upgrade now would be using a more recent
version (e.g. v7.9.x series); however, capturing this error case to help
the odd user in a unique environment.

[1]: https://jira.atlassian.com/browse/CONFSERVER-59196

Signed-off-by: James Knight <james.d.knight@live.com>