initial intersphinx support
The following adds initial support for generating an inventory file to
be used by the `sphinx.ext.intersphinx` extension. This initial release
will build document and various references (e.g. titles) inside an
inventory file.

This file is only built when publishing occurs, as it is the only time
it will know the specific page identifiers to build links for targets.

Signed-off-by: James Knight <james.d.knight@live.com>