build: support both graphviz and inheritance-diagram
This extension originally supported the use of inheritance diagrams via
the `sphinx.ext.inheritance_diagram` extension, and later on supported
more graph capabilities via the `sphinx.ext.graphviz` extension. Both
extensions are handled in the builder which the respective nodes are
converted into images to be included on a Confluence page. When
introducing support for graphviz, inheritance_diagram support broke;
this is a result of the replacement of `graphviz` nodes inside
`inheritance_diagram` nodes.

To correct this issue, `inheritance_diagram` nodes will be always
replaced before attempting to replace `graphviz` nodes. With this
change, the use of either extension in a documentation set will work as
expected.

Signed-off-by: James Knight <james.d.knight@live.com>