storage: support width options in tables
When a table is created, no width hints are provided to the markup. This
allows Confluence to attempt to manage these tables in a flexible
manner. However, reStructuredText does provide a `widths` option for
tables which a user may wish to take advantage of. Since Confluence does
allow width style events to be applied in storage format, attempt to
take advantage of this to help better space columns based off the hints
provided in a documentation's source.

Signed-off-by: James Knight <james.d.knight@live.com>