builder: do not fold various info lines
A series of information lines generated by the builder, where a two
stage message is printed (leading text -> process -> done message),
should have configured the prints to not generate a newline. However,
each of these lines set `nonl` to an incorrect value. This was due to an
original commit [1] introducing the logging pattern and the
implementation being copied over time. This commit finally corrects this
issue.

[1]: 8a39c9d5448b7cca6392a6a167877c0d74b28229

Signed-off-by: James Knight <james.d.knight@live.com>