test: only force index root document for older versions
The testing library will force the default root document to `index` to
help support using a series of standard test documents which can work
on various versions of Sphinx. It is preferred to limit changes to the
configuration unless required; therefore, only adjust the test
configuration when a Sphinx engine less than v2.0 is detected.

Signed-off-by: James Knight <james.d.knight@live.com>