publisher: handle case-insensitive page names
Confluence space names and page names are case-insensitive, but the REST
API is not. This causes issues when a page name is believed to be a new
page and can result in a "title already exists" error when attempting to
published.

A workaround is being introduced when this extension believes it has
detected a case mismatch. In this event, a CQL search is made for page
names with similar names (via the `~` hint). In the event a title match
is detected, it will replace the assumed new page creation to a new page
update.

See also CONFSERVER-40303.

Signed-off-by: James Knight <james.d.knight@live.com>