tests: new validation home tweaks
Atlassian has authorized an open source usage of the organization
`sphinxcontrib-confluencebuilder` [1] on their Cloud environment. This
will be a better demonstration/testing area than the user-specific
location used before (and hopefully, can be used with GitHub actions in
some capacity).

Adjusting the configuration setup for the validation test to take
advantage of this new home. Since this project has an organization, the
validation sets for a given version can now be generated in their own
spaces (without the need to add a version prefix).

[1]: https://sphinxcontrib-confluencebuilder.atlassian.net/

Signed-off-by: James Knight <james.d.knight@live.com>