storage: properly build reftarget download roles
While the `:download:` role helps users link to a non-reST documents, it
can sometimes be used to build a link to a target (observed in the
`html` builder). The original implementation for
`visit_download_reference` built the link, but never properly closed the
tags (the node was skipped instead of processing inwards and eventually
departing); this commit corrects.

Signed-off-by: James Knight <james.d.knight@live.com>