storage: only inject trailing dt tag for last term
When a definition inside a definition includes a term, the
implementation will flag the detection of a term to ensure required
closing flags are populated (as it attempts to process one or more
terms). However, in the event that a following definition entry does not
have a term defined, it will improperly try to pop a non-existent dt
tag. Correct this by unmarking the has-term flag after closing the
previously open term tag.

Signed-off-by: James Knight <james.d.knight@live.com>