handle missing images
While many checks exist to help determine a resource's path, there can
be times where a resource may be unavailable. For example, if an
extension prepares image nodes with paths that do not immediately exist
(e.g. if it assumes that something else will provide them, or the
external extension is experiencing an issue), the node will still need
to be processed by this extension. If the event that a post-build
reftarget/uri points to an unknown asset, generate a warning message and
skip the node (instead of hard failing).

Signed-off-by: James Knight <james.d.knight@live.com>