handle internal document references with refdocname set
Typically, when a reference for a document is found pointing to another
document in the same project, an `internal` flag is set. This is used in
the reference processing to build a macro link to another page in
Confluence. It has recently been observed that some extensions also
build internal document references with a `refdocname` value specified.
This commit tweaks the reference building operation by accepting either
the `internal` flag or `refdocname` attribute as an indication that the
reference is an internal link, and build accordingly.

Signed-off-by: James Knight <james.d.knight@live.com>