ensure allowlist/denylist accepts list from command line
With the rework/introduction of the allowlist/denylist options, the
configuration option permits a string value to pass in a filename (for
a file containing the list of docnames). However, this work broke the
documented support for passing in a list of docnames for the original
`confluence_publish_subset` option.

This commit restores this feature by checking if the provided option is
an override from the command line. If so, the string entry will be split
into list of docname entries.

Signed-off-by: James Knight <james.d.knight@live.com>