builder: set a more specific format
When registered the format for the Confluence builders, the tag value
`confluence` is used as the format identifier. However, to be better
prepared for other formats which may be introduced (such as the
Atlassian document format), it would be ideal to make this a bit more
specific. Adjusting this format to `confluence_storage` to indicate the
Confluence storage format type.

Signed-off-by: James Knight <james.d.knight@live.com>