refactor code of conduct document
This commit moves the code of conduct for this repository out of the
contributing document into its own document (the common
"CODE_OF_CONDUCT" document).

This change also corrects from missing terms from the pledge statement
outlined in the Contributor Covenant v1.4 document as well as adjusts
the contact point to the project's maintainers.

Signed-off-by: James Knight <james.d.knight@live.com>