LICENSE: update most recent license state
This commit makes a series of minor changes to better reflect the
license information for this project. First, the copyright statement has
been updated to match the newer style used in the implementation.
Second, this license has been a 2-clause BSD license, however some of
the formatting/wording was not an exact copy of the original license
text; adjusting to match for consistency. Finally, adding license
information for implementation included in this repository that is not
directly part of the Sphinx Confluence Builder Contributors grouping.

Signed-off-by: James Knight <james.d.knight@live.com>