builder: write empty pages
In the event where a document entry only contains a title reference and
is removed via the `confluence_remove_title` option, the resulting page
will be empty. While this should be fine, the existing builder
implementation would not generate an output document which was empty.
This would leave to an error such as the following:

    WARNING: error reading file<filename>.conf: [Errno 2] No such file or directory: ...

If a documentation set has an empty page, this extension should still be
creating/publishing it. Correcting this in the builder to write content
if even an empty string is set.

Signed-off-by: James Knight <james.d.knight@live.com>