doc: handle version importing in utility scripts
When a user invokes a local documentation build, it will most likely
fail since the import process to extract the version [1] of this
extension cannot be found. Adjusting the respective scripts to work
inside the repository root before attempting to build the module (which
should allow the Python interpreter to load the
`sphinxcontrib.confluencebuilder` import).

[1]: 9e52b399787db80b7ebcbddefd8ae03d5b3de42c

Signed-off-by: James Knight <james.d.knight@live.com>