logger: allow initialization to trigger logging setup
When a Sphinx application starts, it will initialize (among other
things) the logging framework. This includes features like color support
for terminals. It is desired for some upcoming cases to be able to use
this extension's logger instance to generate log messages before a
Sphinx application has been created. When the preload option is set, the
Sphinx logging namespace will have its setup triggered.

Since this setup process requires a Sphinx application instance, which
is not available at this stage. A hack is introduced to create a mock
Sphinx application to allow the setup to continue. Preloading should
only been done for use cases outside the main operation of this
extension.

Signed-off-by: James Knight <james.d.knight@live.com>