introduce a series of mainline actions
This commit introduces a series of mainline actions a user may want to
use. This includes a build, report and wipe actions.

(report)
The report action is the primary reason for the introduction of these
changes. The goal is to provide a means for users to generate a report
of an instance/configuration state which can be captured and shared in
this project's issue tracker. A report will include versions of various
components and include configuration information if working inside a
documentation set.

(wipe)
The wipe action is the second reason for the introduction of mainline
changes. The goal of this action is to provide an easy way for a
developer to wipe the content of a Confluence space. There can be times
where multiple documentation sets are being processed on a space and it
is not always trivial to delete all pages from a space for select
configurations/users. When invoked, this call will use the configured
space as the space to clear. A series of warning/checks will be provided
to help communicate the risk and prevent undesired wipes.

(builder)
This action provide an alternative way for a user to trigger a build
with the Confluence Builder extension. Unknown on if this is a good
long-term usage as it does not use the mainline of the Sphinx
application, but instead loads the Sphinx module and manually invokes
it internally. This is to help pre-load/use common options which most
users may not care to set (output directory, using the `confluence`
builder, etc.).

Signed-off-by: James Knight <james.d.knight@live.com>