tests: simplify doctree output management
The Sphinx application accepts a directory for both output documents and
output doctrees. Initial implementations for unit tests to prepare an
environment for this would always prepare the two directories and pass
them into the application; however, the is not need for any of the
existing unit tests to deal with caring about the location of the
doctrees output location.

Instead, adjust the test utility methods which prepare a Sphinx
application to default any doctree location to a `.doctrees` sub-folder
in a generated output directory (a pattern observed with other stock
Sphinx builders).

Signed-off-by: James Knight <james.d.knight@live.com>