tests: allow sphinx-prep utility methods to auto-generate output dirs
Testing can be cleaned up a bit to permit the automatic generation of an
output directory when preparing/building a Sphinx application. Unit test
names can be taken advantage of to help prepare an output directory
based on inspection. When preparing a Sphinx application or after
invoking a Sphinx build, the output directory can be provided either
through the return on the utility method, or the common output directory
attribute stored in the Sphinx application instance.

Signed-off-by: James Knight <james.d.knight@live.com>