tests: improve build-sphinx explicit filenames handling
A Sphinx engine allows accepting a list of filenames it will process;
however, these filenames need to be set to full paths. This is not
always convenient for testing. This commit changes the call to support
provided relative paths for documents to include.

In addition, when passing a documentation set to process, Sphinx
requires that the documentation set has an existing root document. In
some testing datasets, they may not have one that exists. If this is
detected, this helper will adjust the configuration to adjust the root
document to a provided filename, which should prevent issues when the
Sphinx application prepares an environment.

Signed-off-by: James Knight <james.d.knight@live.com>