tests: adjust test_config to use index over admonitions
The admonitions test logic is about to be moved out of the legacy
container. To prevent issues with this test, do not have it relay on the
admonitions documentation.

This will apply to other related documents it uses in the future, but
that will be addressed at a later time.

Signed-off-by: James Knight <james.d.knight@live.com>