builder: ensure top ref is set for all matching section/title ids
When a top reference is applied to references when a title element is
removed (by default), the initial implementation [1] did not properly
capture all the references to apply the `top-reference` the attribute.
In the common case, the attribute would be applied by chance as the
original code would update the reference the title entry pointed to
(versus a reference pointing to the title/section grouping). However,
this does not properly work for all cases (e.g. when documentation
attempts to use the `:backlinks:` option).

This change includes two fixes when attempting to properly apply an
attribute on related references. First, it finds references which point
to the title to flag for application. Second, the section identifier
(containing the title) will have its identifiers used (as the links may
actually point to the section instead of the contained title).

[1]: 914a6d5c472340fab837e7ef8434935440d7e5e3

Signed-off-by: James Knight <james.d.knight@live.com>