builder: strip underscores for header targets with links
When Confluence generates identifier values for headers, the value is
typically a combination of the page name and the no-space contents of
the title. There appears to be an exception when the header contains a
link to another location (internal or external). For headers containing
links, it has been observed that the generated identifier will remove
all underscores from the identifier value.

When populating targets based off section identifier tags and a pending
reference is detected in the contents of a section, remove any detected
underscore characters.

Signed-off-by: James Knight <james.d.knight@live.com>