tests: output unit test load/import errors when pattern does not match
When a target test pattern is provided and a syntax error exists in the
expected unit test, the unit test helper only indicates that the pattern
cannot be found. A developer most likely assumes the pattern is
incorrect and when it is known to match, the error is being suppressed.

This commit will track any known implementation errors when cycling
through the detected test files. In the event that a pattern is provided
and cannot be found, the first load/import error detected (if any) will
be reported in the output.

Signed-off-by: James Knight <james.d.knight@live.com>