explicit list of support image types
When populating the builders `supported_image_types`, the original
implementation would use a copy of the `StandaloneHTMLBuilder` builder's
supported images [1]. However, Confluence does support a different
subset of image types, so it is most likely better to define an explicit
list which works for both the existing Sphinx engines support and the
target Confluence instances supported (counter to the original review
comments).

This commit introduces `SUPPORTED_IMAGE_TYPES`, a list of mime types
which are registered in the Confluence implementation which also work
with Sphinx. The only change from using `StandaloneHTMLBuilder`s set of
images types to this list, is the added support for BMP files
(`image/x-ms-bmp`).

To provide a bit more flexibility, depending on newer versions of Sphinx
or unique Confluence instance support, the configuration option
`confluence_additional_mime_types` is being introduced to allow users to
extend the supported types if needed.

[1]: https://github.com/sphinx-contrib/confluencebuilder/pull/217

Signed-off-by: James Knight <james.d.knight@live.com>