assets: allow user to force standalone hosting
This extension will automatically attach all assets (images, downloads)
to the root document when detected to be hosted on multiple documents.
This is to minimize the amount of data that needs to be hosted on a
Confluence instance. In some cases, a user may not wish to use this
feature and may desire to host each individual asset for a document in
each of the documents using the respective assets.

This commit introduces the `confluence_asset_force_standalone` option
which forces uploading an asset on each document using said asset.

Signed-off-by: James Knight <james.d.knight@live.com>