config: permit string values for booleans/integers
If a configuration entry is defined as a string (primarily, when
provided via a define in the command line), this extension should be
flexible in accepting the string value which can be converted to the
desired type for a configuration entry. For all boolean values and
integer values, a string value will now be accepted and will be
converted appropriately before the builder attempts to use the any
configuration options.

Signed-off-by: James Knight <james.d.knight@live.com>