config: handle empty strings as unset cases
In the event that a user provides an empty define value for a
configuration entry, treat this as a request to "unset" any existing
configuration value set (i.e. go back to default).

The configuration entry `confluence_secnumber_suffix` is a special case
where an empty string may be desired for the suffix entry; therefore an
empty value via the command line will still result in an empty suffix.

Signed-off-by: James Knight <james.d.knight@live.com>