builder: properly wrap displaymath node contents
The original math rendering only works for single-line contents. When
using a displaymath directive, the content is not properly wrapped and
will always result in a LaTeX processing error. The Sphinx engine
provides a method to help wrap the contents correctly for this node type
(via `wrap_displaymath`); adjusting to use this method.

Signed-off-by: James Knight <james.d.knight@live.com>