support math depth
When a configuration sets `imgmath_use_preview`, the math rendering
request will provide a depth offset which can be used to apply around
math images for improved visual alignment. When a depth is generated
from `render_math`, track this depth into the newly created image node.

For storage format, when an image is assigned with a math-depth value,
include a span wrapper with a vertical alignment offset set.

Signed-off-by: James Knight <james.d.knight@live.com>