assets: support late processing file/image nodes
In the event that a file/image node cannot be pre-processed, visiting
a file/image node in the writer will fail since there is no asset to
build a reference to. This primarily can happen with third-party
extensions which may convert a node type to an image during the writer
phase.

To provide flexibility for these extensions, if an asset is not tracked
when attempting to fetch from the asset management, attempt to process
the node to track a standalone asset.

Signed-off-by: James Knight <james.d.knight@live.com>