provide userspace/password feedback on bad-space error
In the event that a publish attempt generates a bad space error, this is
usually due to a bad space name or authentication issues with a
configuration configured for an anonymous publish attempt. To help
provide a hint that a user may have an anonymous configuration setup,
provide information on the configured username/password state.

Signed-off-by: James Knight <james.d.knight@live.com>