report: indicate if a sensitive configuration is a set empty value
When generating a report of a documentation's configuration state,
help distinguish whether or not a sensitive configuration is an
empty-set value. This is to help provide improve feedback on the
configuration state where string options may be assumed to be set for
some users, but are set to empty strings.

Signed-off-by: James Knight <james.d.knight@live.com>