force docname on singlebuilder asset fetching
When processing runtime generated image nodes, the asset management will
attempt to track newly created assets when preparing a node's
translation. This requires determining the document which the image will
be attached on, which uses the `document` attribute populated through a
Sphinx run. When assembling documents using the `singleconfluence`
builder, this attribute is not provided and injected image nodes will be
ignored. To avoid this, when `singleconfluence` is being invoked,
explicitly pass in the docname to be used to ensure an asset entry can
be fetched.

Signed-off-by: James Knight <james.d.knight@live.com>