builder: fold info lines when in a verbose mode
Status logging on an iterator will not generate newlines to a better
user experience for log message; however, when an instance is configured
for a more verbose logging state, these status messages may be stacked
with verbose messages that may be printed. To avoid this, always set to
fold lines when a verbose mode is set.

Signed-off-by: James Knight <james.d.knight@live.com>