publisher: suppress a failed page delete on missing page id
When purging is enabled, this extension will populate a list of
descendant pages to delete and filter out pages that will get populated
for the new publish set. With the final obsolete list of pages, they
will be marked for deletion. It has been observed in some cases where
Confluence reports descendant page identifiers for pages which no longer
exist (assumed to be cache related). When this extension attempts to
delete the non-existent page, it will generate an API error and stop the
process. Instead of stopping, if a deletion event is reported to not
exist on the Confluence instance, just ignore the error state.

Signed-off-by: James Knight <james.d.knight@live.com>