singlebuilder: register a document target
When building a single Confluence page, if a reference explicitly
references to another document (:doc:<>), there can be no registered
target to properly point to. The `_register_doctree_title_targets` call
focuses on building targets based off the identifiers defined in the
section names; however, there is no guarantee that the assigned
identifiers will match to an internal refid value to a specific document
name (e.g. if "pagea" links to "pageb", page A can have a reference to a
refid "pageb", but page B's root section name *may* not have a matching
identifier).

Since we want references like this to point to the starting location of
these document's content, there is a desire to register a reference to
the first section name value (i.e. first title) if one exists.
Therefore, if the root section name does not register a target which
matches a prospect refid value for a :doc:<> reference, register an
additional target to the leading section which has this mapping.

In addition, when `confluence_remove_title` is set (default), while an
additional target can be registered for the top section, none will exist
for references to the root document if the title is removed. In the
event that this scenario is detected, replace any document references of
the root pages with a `#top` link instead.

Signed-off-by: James Knight <james.d.knight@live.com>