singlebuilder: support lazy doc links for missing titles
Typically, pages will include a title value for a page which can be used
as a target/anchor location when needing to jump to the page's content.
If a page does not have title, this can prevent :doc:<> references from
having a location to jump to when a single doctree is assembled. In the
event that a new document is processed where it detects it has no title
target registered, it will automatically create an anchor (if
permitted), to allow an anchor fallback to have a document reference a
place to jump to.

Signed-off-by: James Knight <james.d.knight@live.com>