do not override docutils node-visitor's document attrib
This extension's base translator (`ConfluenceBaseTranslator`) is based
off of docutil's `NodeVisitor` implementation. When a document finally
departs, it would populate a `document` attribute which the writer
implementation could fetch before writing content to a file. However,
`NodeVisitor` already uses a `document` attribute for other purposes,
and overriding the attribute causes issues for Sphinx/docutils calls
which use this attribute. Instead, renaming this attribute to
`body_final`.

Signed-off-by: James Knight <james.d.knight@live.com>