handle toctree caption changing from nodes.caption to nodes.title
As of Sphinx v3.5.4, changes have been made to the application where a
`:caption:` hint on a toctree node will create a `nodes.title` instance
instead of an `nodes.caption` instance. This modification was introduced
to support styling changes introduced in docutils-0.17 [1].

This commit adjusts the `visit_title` to detect if the title entry is
being built from a toctree node -- if so, forward the node processing
back to the existing caption implementation.

[1]: sphinx-doc/sphinx#9053

Signed-off-by: James Knight <james.d.knight@live.com>