doc: promote ask-password
Adjusting the documentation to promote the `confluence_ask_password`
option over `confluence_server_pass`. User feedback has mentioned
multiple times that the documentation appears to be promoting the
storing of secret in shared repositories. To help alleviate these
concerns (in part with changes [1] to `confluence_ask_password` to
better support MSYS/MinGW environments), areas of the documentation
which hint to specify a password value will be replaced with the
ask/prompt option. Users who want to explicitly set password options
will find the options when browsing the documentation and observe the
cautions associated with the configurations (which may help reduce users
from storing secrets in a shared repository).

[1]: 20ac6a9ee026021eada92bc0d2b820a840b1fe58

Signed-off-by: James Knight <james.d.knight@live.com>