storage: support emphasis of identifier values in signatures
For releases of Sphinx pre-v4.x, signature descriptors could contain
identifier entries wrapped in an emphasis node. In the most recent
version of Sphinx (observed in 4.0.0b2), identifiers may not be wrapped
in an emphasis node. Instead of relying on an emphasis node, use the
class hint which identifiers an identifiers type [1] to build an `em`
container.

[1]: https://github.com/sphinx-doc/sphinx/blob/v3.5.4/sphinx/addnodes.py#L224-L226

Signed-off-by: James Knight <james.d.knight@live.com>