doc: only border/color literal-marked pre-span
Documentation will style `pre` class `span` entries with a
border/color/etc. to help promote/identify the literals through the
documentation set. It has been noticed that `confval` entries in newer
Sphinx builds will be assigned the same style, which is not desired to
be added to these configuration title entries.

Adjusting the styling to only apply the style modifications if the
pre-span type is contained inside a literal marked code element.

Signed-off-by: James Knight <james.d.knight@live.com>