support raw html injection
This commit provides an advanced option to allow raw html injection in
the output of a storage-format generated document. This is to help users
wanting to somewhat support raw HTML content generated from Markdown
sources.

This is not always guaranteed to work; the raw html content may not be
compatible with Atlassian's storage format. Results may fail to publish
or contents may be suppressed on the Confluence instance.

Signed-off-by: James Knight <james.d.knight@live.com>