storage: only apply default alignment on figure entities
With a previous change helping manage new alignment hints in Sphinx [1],
these changes would apply default alignment hints to nodes which should
not be using them. For example, table of contents captions and raw
images (outside of figures) would apply a default alignment of `center`,
where no implicit alignment is applied when using the `html` builder.

To correct this, only attempt to apply alignments to captions, legends
and images if their parent is a figure node.

[1]: 50fd9ca636c54ab8fb272fe489154a3dcdf0ca73

Signed-off-by: James Knight <james.d.knight@live.com>