adjust attachment no-check header to most recent value
Based off some newer documentation and implementation checks, it appears
that no-check value to be set for the `X-Atlassian-Token` header should
be a value of `no-check` instead of `nocheck`; adjusting.

This commit also adds an undocumented `CONFLUENCEBUILDER_LEGACY_NOCHECK`
environment option to use the legacy header value for users who may
experience issues in their environments.

Signed-off-by: James Knight <james.d.knight@live.com>