support cell alignment hints injected in the classes attribute
It has been observed that some extensions (e.g. MyST-Parser) will inject
alignment hints for cells into a node's `classes` attribute. This would
allow a styling to define a, for example, `text-align:center` class
which can include a center alignment hint for formatting on a document.

In the event that a text-align style has been detected inside a class
list on a node, directly inject it into the style attribute of the
paragraph holding the cell data.

Signed-off-by: James Knight <james.d.knight@live.com>