only inject sphinx.ext.imgmath if confluence building
This extension would automatically inject the `sphinx.ext.imgmath`
extension into the Sphinx application if the extension was not detected
as configured by a user. This was to help users easily use the only
supported math-handling extension for math directives when publishing to
Confluence targets.

An issue with this approach is that this would also inject the extension
for a Sphinx invoke, even when not using a Confluence builder (e.g. if
a user is building an `html` builder). This would override any other
math extension (e.g. `sphinx.ext.jsmath`) that may be registered. To
prevent this, only lazily inject `sphinx.ext.imgmath` when using a
Confluence-based builder.

Signed-off-by: James Knight <james.d.knight@live.com>