publisher: force v1 editor
The Sphinx Confluence Builder still only supports the legacy display
with storage format content generated from the storage format
translator. Recent changes show that even when publishing with a
representation request of `storage`, the content will be accepted but a
newer editor may be used. This is undesired at this time (until an ADF
translator is introduced).

To ensure we use the legacy editor, inject the metadata for new pages
to explicitly indicate that a `v1` editor must be used. This appears to
work with Confluence Cloud and (at least) the oldest supported
Confluence Server revision. Note that the metadata is not applied on
updated pages as the forced conversion is not supported by an update
request.

Signed-off-by: James Knight <james.d.knight@live.com>