README.rst: direct link to stable verification page
Most users who may stumble on the README looking for a demonstration
will most likely not care about the specific contents of the validation
documents (and more so on what Confluence looks like). Dropping the
leading portion of the demonstration section to no longer list the
source link.

In addition, the generic landing page for the Confluence Cloud space
(sphinxcontrib-confluencebuilder) can list multiple spaces which a user
may also not care about. Instead, directly link to the most recent
stable validation set providing during a release.

Signed-off-by: James Knight <james.d.knight@live.com>