refactor confluence builder initialization tasks
This commit reworks a significant porition of this extension's
initialization work into a single `builder-inited` event. In most cases,
several of the initialization events such as asset processing, imgmath
injection, etc. were already managed through their own individual
events. Now, only a single hook is made where all these actions can be
invoked (still specific to this extension's builder types only).

In addition, this commit also moves the registering of this extensions
directives in this event as well. It is most likely that all cases
where this extension-specific directives are used are solely for this
extensions use. By registering them after the builder is initialized
(where we can check which builder is being used -- this extension type
or some other standard/third-party builder), we avoid triggering
builder errors for documentation which include the use of these
directives in the documentation. Specifically, instead of having a
configured translator failing on processing an unknown node type, users
will be presented with a warning that the directive type is unknown (and
continue building).

Signed-off-by: James Knight <james.d.knight@live.com>