directives-jira: improve non-set jira server handling
Improve handling when a JIRA directive specifies a server but no server
entries are provided in the configuring. Specifically, the option
`confluence_jira_servers` will exists in the configuration (a "default"
`None` value) and may not be set to a dictionary. This will cause the
directive processing to fail when attempting to cycle on the
configuration's value. Instead, adjust the check to ignore both an
unset/empty value.

Signed-off-by: James Knight <james.d.knight@live.com>